nova rota criada e rota "random users" modificada

+++ Routes/Post.js
@@ -1,7 +1,6 @@
 import express from "express";
 import { User, Post, Comment, sequelize } from "../model/index.js";
 import verifytoken from "../middlewares/verifytoken.js";
-import { Op } from "sequelize";
 
 const postRouter = express.Router();
 
@@ -9,7 +8,6 @@ const postRouter = express.Router();
 
 postRouter.post("/post", verifytoken, async (req, res) => {
   const id = req.id;
-
   const { img, legend } = req.body;
 
   if (!id) {
@@ -18,17 +16,15 @@ postRouter.post("/post", verifytoken, async (req, res) => {
       .json({ message: "Não foi possivel fazer a publicação" });
   }
 
-  if (!img) {
-    return res.status(400).json({ message: "poste uma imagem" });
-  }
-
-  if (!legend) {
-    return res.status(400).json({ message: "digite uma legenda" });
+  if (legend.length > 500) {
+    return res
+      .status(400)
+      .json({ messge: "A legenda da publicação é muito grande" });
   }
 
   const post = await Post.create({
-    img_post: img,
-    legend: legend,
+    img_post: img && img,
+    legend: legend && legend,
     userId: id,
   });
 
@@ -46,13 +42,6 @@ postRouter.post("/post", verifytoken, async (req, res) => {
 // ------------------- ALL POSTS ------------------- //
 
 postRouter.get("/post", async (req, res) => {
-  // const posts = await Post.destroy({
-  //   where: {
-  //     id: {
-  //       [Op.gt]: 10,
-  //     },
-  //   },
-  // });
   const posts = await Post.findAll({
     attributes: ["id", "img_post", "legend", "createdAt"],
     order: [["createdAt", "DESC"]],
@@ -63,87 +52,16 @@ postRouter.get("/post", async (req, res) => {
       },
       {
         model: Comment,
-        where: {},
         attributes: ["id", "content", "createdAt"],
         include: {
           model: User,
-          attributes: ["id", "avatar", "username", "createdAt"],
+          attributes: ["id", "avatar", "username"],
         },
         separate: true,
         order: [["createdAt", "DESC"]],
       },
     ],
   });
-
-  // const post = await sequelize.query(
-  //   'SELECT "Posts"."id", "Posts"."img_post", "Posts"."legend", "Posts"."createdAt", "User"."id" AS "User.id", "User"."avatar" AS "User.avatar", "User"."username" AS "User.username" FROM "Posts" AS "Posts" LEFT OUTER JOIN "Users" AS "User" ON "Posts"."userId" = "User"."id" ORDER BY "Posts"."createdAt" DESC;',
-  //   'SELECT "Comment"."id", "Comment"."content", "Comment"."createdAt", "Comment"."postId", "User"."id" AS "User.id", "User"."avatar" AS "User.avatar", "User"."username" AS "User.username", "User"."createdAt" AS "User.createdAt" FROM "Comments" AS "Comment" LEFT OUTER JOIN "Users" AS "User" ON "Comment"."userId" = "User"."id" ORDER BY "Comment"."createdAt" DESC LIMIT 1;',
-  //   { type: sequelize.QueryTypes.SELECT }
-  // );
-
-  // const Id = req.id;
-
-  // if (!Id) {
-  //   return res.status(401).json({
-  //     message: "acesso negado, token inválido, não retornou id",
-  //   });
-  // }
-  // // trazer quem eu estou seguindo, tentar usar ids dessa busca na busca dos posts
-  // const following = await User.findByPk(Id, {
-  //   attributes: ["id", "avatar", "username", "name"],
-  //   include: [
-  //     {
-  //       model: Following,
-  //       include: [
-  //         {
-  //           model: User,
-  //           attributes: ["id", "avatar", "username"],
-  //         },
-  //       ],
-  //     },
-  //   ],
-  // });
-
-  // // trazer todos os posts do usuario e de quem ele segui
-
-  // const followingMap = following.Followings.map((user) => user.follow);
-
-  // await followingMap.push(Id);
-
-  // const posts = await Post.findAll({
-  //   where: {
-  //     userId: followingMap,
-  //   },
-  // attributes: ["id", "img_post", "legend", "createdAt"],
-  // order: [["createdAt", "DESC"]],
-  // include: [
-  //   {
-  //     model: User,
-  //     attributes: ["id", "avatar", "username"],
-  //   },
-  //   {
-  //     model: Comment,
-  //     attributes: ["id", "content", "createdAt"],
-  //     include: {
-  //       model: User,
-  //       attributes: ["id", "avatar", "username", "createdAt"],
-  //     },
-  //     separate: true,
-  //     order: [["createdAt", "DESC"]],
-  //   },
-  // ],
-  // });
-
-  // const { id, img_post, legend, createdAt, 'User.id', 'User.avatar',
-  // 'User.username' } = post;
-
-  // const newModelPosts = {
-  //   id: id,
-  //   img_post: img_post,
-  //   legend: legend,
-  //   createdAt: createdAt,
-  //   comment: post,
-  // };
 
   if (!posts) {
     return res.status(400).json({ message: "postagens não encontradas" });
@@ -191,7 +109,7 @@ postRouter.get("/post/:postId", async (req, res) => {
     return res.status(400).json({ message: "publicação não encontrada" });
   }
 
-  const posts = await Post.findAll({
+  const post = await Post.findOne({
     where: {
       id: postId,
     },
@@ -205,7 +123,7 @@ postRouter.get("/post/:postId", async (req, res) => {
         attributes: ["id", "content", "createdAt"],
         include: {
           model: User,
-          attributes: ["id", "avatar", "username", "createdAt"],
+          attributes: ["id", "avatar", "username"],
         },
         separate: true,
         order: [["createdAt", "DESC"]],
@@ -214,16 +132,73 @@ postRouter.get("/post/:postId", async (req, res) => {
     attributes: ["id", "img_post", "legend", "createdAt"],
   });
 
-  if (Boolean(posts === false)) {
+  if (!post || post === [] || post === null || post.lenght < 1) {
     return res.status(400).json({ message: "postagens não encontradas" });
   }
 
   try {
-    return res.status(200).json(posts);
+    return res.status(200).json(post);
   } catch (error) {
     return res
       .status(500)
       .json({ message: "ocrreu um erro, tente mais tarde", error });
+  }
+});
+
+//                                                      //
+// -------------------- PUT POST ------------------- //
+//                                                      //
+
+postRouter.put("/post/:postId", verifytoken, async (req, res) => {
+  const postId = req.params.postId;
+  const userId = req.id;
+
+  const { ImgUrl, legend } = req.body;
+
+  console.log(ImgUrl, legend);
+
+  if (!userId) {
+    return res.status(400).json({ message: "usuario não encontrado" });
+  }
+  if (!postId) {
+    return res.status(400).json({ message: "publicação não encontrada" });
+  }
+
+  if (legend.length > 500) {
+    return res
+      .status(400)
+      .json({ messge: "A legenda da publicação é muito grande" });
+  }
+
+  const verify = await Post.findOne({ where: { id: postId, userId: userId } });
+
+  if (!verify) {
+    return res.status(400).json({ message: "você não fez essa publicação" });
+  }
+
+  const post = await Post.update(
+    {
+      img_post: ImgUrl && ImgUrl,
+      legend: legend && legend,
+    },
+    {
+      where: { userId: userId, id: postId },
+    }
+  );
+
+  if (!post) {
+    console.log(post);
+    return res
+      .status(400)
+      .json({ message: "não foi possivel atulizar sua publicação" });
+  }
+
+  try {
+    return res
+      .status(201)
+      .json({ message: "publicação atualizada", reaload: "true" });
+  } catch (error) {
+    return res.status(500).json(error);
   }
 });
 
@@ -240,6 +215,12 @@ postRouter.delete("/post/:postId", verifytoken, async (req, res) => {
   }
   if (!postId) {
     return res.status(400).json({ message: "publicação não encontrada" });
+  }
+
+  const verify = await Post.findOne({ where: { id: postId, userId: userId } });
+
+  if (!verify) {
+    return res.status(400).json({ message: "você não fez essa publicação" });
   }
 
   const user = await Post.destroy({
