+++ Routes/following.js
@@ -1,20 +1,33 @@
 import express from "express";
 import verifytoken from "../middlewares/verifytoken.js";
-import { Following } from "../model/index.js";
+import { Followers } from "../model/Followers.js";
+import { Following, User } from "../model/index.js";
 
-export const followingRouter = express.Router();
+const followingRouter = express.Router();
 
 //                                                      //
 // --------------------- CREATE ----------------------- //
 //                                                      //
-
-followingRouter.post("/following/:followId", verifytoken, async (req, res) => {
+followingRouter.post("/follow/:followId", verifytoken, async (req, res) => {
   const userId = req.id;
   const followId = req.params.followId;
 
+  if (!userId) {
+    return res.status(403).json({ message: "seu ID não foi encontrado" });
+  }
+
+  if (!followId) {
+    return res.status(403).json({ message: "followId não encontrado" });
+  }
+
   const follow = await Following.create({
     userId: userId,
-    follow: followId,
+    followingId: followId,
+  }).then(() => {
+    Followers.create({
+      followersId: userId,
+      userId: followId,
+    });
   });
 
   try {
@@ -26,25 +39,24 @@ followingRouter.post("/following/:followId", verifytoken, async (req, res) => {
   }
 });
 
-followingRouter.get("/following/:followId", async (req, res) => {
+//                                                      //
+// --------------------- GET ONE ---------------------- //
+//                                                      //
+
+followingRouter.get("/user/:userId/following/:followId", async (req, res) => {
+  const userId = req.params.userId;
   const followId = req.params.followId;
 
-  const followers = await Following.findAll({
-    // where: {
-    //   follow: followId,
-    // },
+  const following = await Following.findOne({
+    where: {
+      followingId: followId,
+      userId: userId,
+    },
+    include: { model: User, attributes: ["name", "email", "username"] },
   });
 
-  // const followersId = followers.map((users) => users.userId);
-
-  // const followersResult = await User.findAll({
-  //   where: {
-  //     id: followersId,
-  //   },
-  // });
-
   try {
-    return res.status(201).json(followers);
+    return res.status(201).json(following);
   } catch (error) {
     return res
       .status(500)
@@ -52,12 +64,39 @@ followingRouter.get("/following/:followId", async (req, res) => {
   }
 });
 
+//                                                      //
+// --------------------- GET ALL ---------------------- //
+//                                                      //
+
+followingRouter.get("/user/:userId/following", async (req, res) => {
+  const userId = req.params.userId;
+
+  const following = await Following.findAll({
+    where: { userId: userId },
+    include: {
+      model: User,
+      attributes: ["id", "name", "email", "username", "avatar"],
+    },
+  });
+
+  try {
+    return res.status(200).json(following);
+  } catch (error) {
+    return res
+      .status(500)
+      .json({ message: "Desculpe ocoreu um erro, tente mais tarder", error });
+  }
+});
+
+//                                                      //
+// --------------------- DELETE FOLLOWING ---------------------- //
+//                                                      //
+
 followingRouter.delete(
   "/following/:followId",
   verifytoken,
   async (req, res) => {
     const userId = req.id;
-
     const followId = req.params.followId;
 
     if (!userId) {
@@ -72,10 +111,28 @@ followingRouter.delete(
       });
     }
 
-    Following.destroy({
+    const verifyRegister = await Following.findOne({
       where: {
         userId: userId,
-        follow: followId,
+        followingId: followId,
+      },
+    });
+
+    if (!verifyRegister) {
+      return res.status(403).json({ message: "você não segue este usuário" });
+    }
+
+    await Following.destroy({
+      where: {
+        userId: userId,
+        followingId: followId,
+      },
+    });
+
+    await Followers.destroy({
+      where: {
+        userId: followId,
+        followersId: userId,
       },
     })
       .then(() => {
@@ -87,4 +144,4 @@ followingRouter.delete(
   }
 );
 
-// export default followingRouter;
+export default followingRouter;
